+++ CHANGELOG.md
@@ -9,7 +9,8 @@ with respect to its command line interface and HTTP interface.
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.57...HEAD)
 * Server: HTTP client requests to Singularity log to Kafka
 * Server: HTTP client requests to the Docker registry log to Kafka
-* [TODO] Server: HTTP server responses log to Kafka
+* Server: HTTP server responses log to Kafka
+* [TODO] Server: Additional metrics reported per request kind, per host and per kind,host pair.
 
 ## [0.5.57](//github.com/opentable/sous/compare/0.5.56...0.5.57)
 ### Fixed